+++ src/app/static/src/app/store/load/actions.ts
@@ -2,15 +2,19 @@ import {ActionContext, ActionTree} from "vuex";
 import {LoadingState, LoadState} from "./state";
 import {emptyState, RootState} from "../../root";
 import {api} from "../../apiService";
-import {constructRehydrateProjectState, flatMapControlSection, verifyCheckSum} from "../../utils";
+import {
+    constructRehydrateProjectState,
+    flatMapControlSections,
+    verifyCheckSum
+} from "../../utils";
 import {Dict, LocalSessionFile, VersionDetails} from "../../types";
 import {localStorageManager} from "../../localStorageManager";
 import {router} from "../../router";
 import {currentHintVersion} from "../../hintVersion";
 import {initialStepperState} from "../stepper/stepper";
 import {ModelStatusResponse, ProjectRehydrateResultResponse} from "../../generated";
-import {ModelCalibrateState} from "../modelCalibrate/modelCalibrate";
 import {DynamicFormData} from "@reside-ic/vue-dynamic-form";
+import {ModelCalibrateState} from "../modelCalibrate/modelCalibrate";
 
 export type LoadActionTypes = "SettingFiles" | "UpdatingState" | "LoadSucceeded" | "ClearLoadError" | "PreparingRehydrate" | "SaveProjectName" | "RehydrateStatusUpdated" | "RehydratePollingStarted" | "RehydrateResult" | "SetProjectName" | "RehydrateCancel"
 export type LoadErrorActionTypes = "LoadFailed" | "RehydrateResultError"
@@ -141,7 +145,7 @@ const getRehydrateResult = async (context: ActionContext<LoadState, RootState>) 
         .get<ProjectRehydrateResultResponse>(`rehydrate/result/${rehydrateId}`);
 
     if (response && response.data) {
-        const {files, savedState} = constructRehydrateProjectState(rootState, response.data)
+        const {files, savedState} = await constructRehydrateProjectState(context, response.data)
 
         if (!rootGetters.isGuest) {
             await dispatch("projects/createProject",
@@ -157,6 +161,7 @@ const getRehydrateResult = async (context: ActionContext<LoadState, RootState>) 
 
             Object.assign(rootState, newRootState);
         }
+
 
         await getFilesAndLoad(context, files, savedState)
     }
@@ -195,7 +200,7 @@ async function getFilesAndLoad(context: ActionContext<LoadState, RootState>,
 // getCalibrateOptions extracts calibrate options from Dynamic Form, this allows
 // backward compatibility supports for calibrate option bug
 const getCalibrateOptions = (modelCalibrate: ModelCalibrateState): DynamicFormData => {
-    const allControlGroups = flatMapControlSection(modelCalibrate.optionsFormMeta.controlSections);
+    const allControlGroups = flatMapControlSections(modelCalibrate.optionsFormMeta.controlSections);
     return allControlGroups.reduce<DynamicFormData>((options, option): DynamicFormData => {
         option.controls.forEach(option => {
             options[option.name] = option.value || null